+++ src/views/helpers/index.ts
@@ -1,1 +1,2 @@
 export { resolveWorkspaceName } from "./workspaceResolver";
+export { getCurrentWorkspaceRepo, parseOriginUrl } from "./currentRepo";