+++ src/views/index.ts
@@ -4,4 +4,5 @@ export { StatusBarIndicator } from "./statusBar";
 export { DashboardDataAssembler } from "./dashboardDataAssembler";
 export type { DashboardRawData } from "../shared/dashboardTypes";
 export { ContextTracker } from "./contextTracker";
+export { getCurrentWorkspaceRepo } from "./helpers";
 export { simplifyModelName, formatDuration } from "./treeViewFormatting";